Move branding to dedicated pkg
